+++ dz21/pom.xml
@@ -60,7 +60,7 @@
         <dependency>
             <groupId>org.postgresql</groupId>
             <artifactId>postgresql</artifactId>
-            <version>42.2.4</version>
+            <version>42.2.5</version>
             <scope>runtime</scope>
         </dependency>
         <dependency>
@@ -70,19 +70,23 @@
         <dependency>
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-acl</artifactId>
+            <!--<version>5.1.1.RELEASE</version>-->
         </dependency>
         <dependency>
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-config</artifactId>
+            <!--<version>5.1.1.RELEASE</version>-->
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context-support</artifactId>
+            <!--<version>5.1.1.RELEASE</version>-->
         </dependency>
         <dependency>
             <groupId>net.sf.ehcache</groupId>
-            <artifactId>ehcache-core</artifactId>
-            <version>2.6.11</version>
+            <artifactId>ehcache</artifactId>
+            <!--<artifactId>ehcache-core</artifactId>-->
+            <!--<version>2.6.11</version>-->
         </dependency>
         <dependency>
             <groupId>org.springframework.security</groupId>
