+++ dz21/src/main/java/ru/otus/dz21/security/UserRoleEnum.java
@@ -1,0 +1,7 @@
+package ru.otus.dz21.security;
+
+public enum UserRoleEnum {
+    ADMIN,
+    ADVANCED_USER,
+    USER
+}
