+++ dz21/src/main/resources/data.sql
@@ -17,11 +17,11 @@ INSERT INTO acl_sid (id, principal, sid) VALUES
   (3, FALSE, 'ROLE_ADVANCED_USER');
 
 INSERT INTO acl_class (id, class) VALUES
+--   (1, 'ru.otus.dz21.domain.Book.class');
   (1, 'ru.otus.dz21.domain.Book');
 
 INSERT INTO acl_object_identity
-  (id, object_id_class, object_id_identity,
-  parent_object, owner_sid, entries_inheriting)
+  (id, object_id_class, object_id_identity, parent_object, owner_sid, entries_inheriting)
   VALUES
   (1, 1, 1, NULL, 1, FALSE ),
   (2, 1, 2, NULL, 1, FALSE ),
@@ -29,14 +29,13 @@ INSERT INTO acl_object_identity
   (4, 1, 4, NULL, 1, FALSE );
 
 INSERT INTO acl_entry
-  (id, acl_object_identity, ace_order,
-  sid, mask, granting, audit_success, audit_failure)
+  (id, acl_object_identity, ace_order, sid, mask, granting, audit_success, audit_failure)
   VALUES
   (1, 1, 1, 2, 1, TRUE , TRUE , TRUE ),
-  (2, 1, 2, 3, 1, TRUE , TRUE , TRUE ),
+  (2, 1, 2, 3, 2, TRUE , TRUE , TRUE ),
   (3, 2, 1, 2, 1, TRUE , TRUE , TRUE ),
-  (4, 2, 2, 3, 1, TRUE , TRUE , TRUE ),
+  (4, 2, 2, 3, 2, TRUE , TRUE , TRUE ),
   (5, 3, 1, 2, 1, TRUE , TRUE , TRUE ),
-  (6, 3, 2, 3, 1, TRUE , TRUE , TRUE ),
-  (7, 4, 1, 2, 1, TRUE , TRUE , TRUE );
-  (8, 4, 2, 3, 1, TRUE , TRUE , TRUE );
+  (6, 3, 2, 3, 2, TRUE , TRUE , TRUE ),
+  (7, 4, 1, 2, 1, TRUE , TRUE , TRUE ),
+  (8, 4, 2, 3, 2, TRUE , TRUE , TRUE );
