+++ dz21/src/main/resources/templates/error.html
@@ -4,6 +4,6 @@
     <meta charset="UTF-8"/>
 </head>
 <body>
-<h1>Доступ запрещён!</h1>
+<h1>Error</h1>
 </body>
 </html>
